% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-app-style.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add a tooltip to an element}
\usage{
add_tooltip(
  id,
  tooltip,
  placement = "right",
  allowHTML = TRUE,
  use_id_only = FALSE,
  ...
)
}
\arguments{
\item{id}{The id of the element to add the tooltip to.}

\item{tooltip}{The tooltip to add.}

\item{placement}{The placement of the tooltip.}

\item{allowHTML}{Whether to allow HTML in the tooltip.}

\item{use_id_only}{Whether to use the id only. Default is \code{FALSE} which
will add the namespace 'tooltip-icon' to the id. Set to \code{TRUE} if the
tooltip icon was not added with \code{add_tooltip_icon} (e.g., if
the tooltip icon was added manually).}

\item{...}{Arguments passed to \code{tippy_this}.}
}
\value{
The element with the tooltip.
}
\description{
Add a tooltip to an element
}
\keyword{internal}
