% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{apply_post_rf_preprocessing}
\alias{apply_post_rf_preprocessing}
\title{Postprocessing helper functions after fitting the RF in NeRF+ models}
\usage{
apply_post_rf_preprocessing(object, x)
}
\arguments{
\item{object}{A \code{ranger} object that has been fitted to the data.}

\item{x}{A data frame or matrix containing the data to be processed.}
}
\value{
A data frame or matrix with the same structure as \code{x}, but with
factor levels converted to numeric values based on the levels stored in
\code{object}.
}
\description{
This function applies some processing steps to the data after
fitting the RF in NeRF+ models. Specifically, it converts factor levels
to numeric values based on the levels stored in the RF model.
}
\keyword{internal}
