% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-app-style.R
\name{display_inline}
\alias{display_inline}
\title{Display inline block}
\usage{
display_inline(obj)
}
\arguments{
\item{obj}{Object to display inline.}
}
\description{
Display inline block
}
\keyword{internal}
