% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{dummy_code}
\alias{dummy_code}
\alias{fit_dummy_code}
\alias{apply_dummy_code}
\title{Dummy code categorical features}
\usage{
fit_dummy_code(x)

apply_dummy_code(dummy_fit, x)
}
\arguments{
\item{x}{A data frame or matrix containing the data to be dummy-coded.}

\item{dummy_fit}{A previously fitted dummy coding model; typically the
output of \code{fit_dummy_code()}.}
}
\value{
For \code{fit_dummy_code()}, a list containing:
\itemize{
\item \code{dummy_fit}: The fitted dummy coding model, which can be used to
dummy-code new data.
\item \code{x}: A data frame or matrix containing the dummy-coded data.
For \code{apply_dummy_code()}, a data frame or matrix containing the
dummy-coded data.
}
}
\description{
Dummy code categorical features
}
\keyword{internal}
