% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-app-modules.R
\name{fileInputUI}
\alias{fileInputUI}
\alias{fileInputServer}
\title{File input module - UI}
\usage{
fileInputUI(
  id,
  label = "File Upload",
  accept = c(".csv", ".txt", ".rds"),
  show_accept = TRUE,
  tooltip = NULL
)

fileInputServer(id, default_data = NULL)
}
\description{
File input module - UI

File input module - Server
}
\keyword{internal}
