% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-embeddings.R
\name{fit_network_embedding}
\alias{fit_network_embedding}
\title{Wrapper to fit various network embedding methods}
\usage{
fit_network_embedding(
  A,
  ndim = 2,
  method = c("adjacency", "laplacian", "score"),
  regularization = 0.5,
  varimax_rotation = FALSE
)
}
\arguments{
\item{A}{Adjacency matrix.}

\item{ndim}{Number of dimensions in embedding.}

\item{method}{Embedding type(s), at least one of "adjacency", "laplacian", or
score".}

\item{regularization}{Regularization parameter for the adjacency matrix.}

\item{varimax_rotation}{Whether to apply varimax rotation to the embedding.}
}
\value{
A list containing the following components:
\itemize{
\item \code{X}: A list of matrices, each corresponding to a method in \code{method}, where
each matrix has \code{ndim} columns representing the embedding.
\item \code{eval_ls}: A list of eigenvalues corresponding to each method in \code{method}.
\item \code{method}: The method(s) used for embedding.
\item \code{regularization}: The regularization parameter used for the adjacency
matrix.
}
}
\description{
Wrapper to fit various network embedding methods
}
\examples{
data(example_data)
embedding_fit <- fit_network_embedding(
  example_data$A, ndim = 2, method = "laplacian"
)

}
