% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{interpret_nerfplus}
\alias{interpret_nerfplus}
\title{Interpret NeRF+ Model}
\usage{
interpret_nerfplus(
  object,
  x,
  x_embed = NULL,
  y,
  A,
  nodeids = NULL,
  xtest,
  xtest_embed = NULL,
  ytest,
  A_full,
  nodeids_test = NULL,
  methods = c("permute", "mdi+", "local", "loo"),
  B = 10,
  metric = NULL,
  save = FALSE,
  save_dir = "results_nerfplus"
)
}
\arguments{
\item{object}{A fitted NeRF+ model object.}

\item{x}{A data frame or matrix containing the training data.}

\item{x_embed}{An optional data frame or matrix of network embeddings corresponding
to the training samples. Only needed if training embeddings were manually inputted.}

\item{y}{A vector of responses for the training data.}

\item{A}{An adjacency matrix representing the network structure for the training nodes.}

\item{nodeids}{An optional vector of node IDs of length n. If provided, node IDs
indicate the rows of A corresponding to each sample. If not provided, the rows
of A are assumed to be in the same order as the rows of x and y.}

\item{xtest}{An optional data frame or matrix containing the test data.}

\item{xtest_embed}{An optional data frame or matrix of network embeddings corresponding
to the test samples.}

\item{ytest}{An optional vector of responses for the test data.}

\item{A_full}{An optional adjacency matrix representing the network structure for
the full set of nodes (training + testing nodes in that order).}

\item{nodeids_test}{An optional vector of node IDs for the test data.}

\item{methods}{A character vector of methods to run. Can include any of
"permute", "mdi+", "local", and "loo".}

\item{B}{Number of permutations to use for permutation feature importance.}

\item{metric}{A function to compute the metric used for global feature
importances. Defaults to R-squared for regression and AUROC for
classification.}

\item{save}{Whether to save the results to disk.}

\item{save_dir}{Directory to save results to if \code{save = TRUE}.}
}
\value{
A list with the following components:
\item{object}{The fitted NeRF+ model.}
\item{data_list}{A list containing the data used for interpretation.}
\item{fi_results}{A list containing the feature importance results.}
\item{loo_results}{A list containing the sample influence results.}
}
\description{
Wrapper function around \code{get_feature_importances()} and
\code{get_loo()}. Helpful to format results correctly for input into NeRF+
Interpreter Shiny application.
}
\examples{
\donttest{
data(example_data)
nerfplus_out <- nerfplus(
  x = example_data$x, y = example_data$y, A = example_data$A,
  lambda_netcoh = 1,
  lambda_embed = 0.1,
  lambda_raw = 2,
  lambda_stump = 3,
  family = "linear", embedding = "laplacian", sample_split = "none"
)
out <- interpret_nerfplus(
  nerfplus_out, x = example_data$x, y = example_data$y, A = example_data$A,
  xtest = example_data$xtest, ytest = example_data$ytest,
  A_full = example_data$A_full
)
}

}
