% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{make_laplacian}
\alias{make_laplacian}
\title{Construct a dense graph Laplacian with optional ridge adjustment}
\usage{
make_laplacian(A, lambda_l = 0, nodedegrees = NULL)
}
\description{
Construct a dense graph Laplacian with optional ridge adjustment
}
\keyword{internal}
