% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{make_sparse_laplacian}
\alias{make_sparse_laplacian}
\title{Construct a sparse graph Laplacian with optional ridge adjustment}
\usage{
make_sparse_laplacian(A, lambda_l = 0, nodedegrees = NULL)
}
\description{
Construct a sparse graph Laplacian with optional ridge adjustment
}
\keyword{internal}
