% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{nerfplus_augmentation}
\alias{nerfplus_augmentation}
\alias{fit_augmentation}
\alias{apply_augmentation}
\title{Augment Psi matrix in NeRF+ with raw features}
\usage{
fit_augmentation(x, psi, tree_info = NULL, include_raw = TRUE)

apply_augmentation(
  x,
  psi,
  tree_info = NULL,
  include_raw = TRUE,
  dummy_fit = NULL
)
}
\arguments{
\item{x}{A data frame or matrix containing the raw features.}

\item{psi}{A data frame or matrix containing the Psi matrix, where each
column corresponds to a decision stump feature.}

\item{tree_info}{Output of \code{\link[ranger:treeInfo]{ranger::treeInfo()}} for a single tree.}

\item{include_raw}{Logical indicating whether to include the raw covariates
in the NeRF+ model. Default is \code{TRUE}.}

\item{dummy_fit}{A previously fitted dummy coding model; typically the output
of \code{fit_dummy_code()}. If \code{include_raw = TRUE}, this is used to dummy-code
raw features.}
}
\value{
For \code{fit_augmentation()}, a list containing:
\itemize{
\item \code{x}: A matrix containing the augmented data, which includes both the raw
features (if \code{include_raw = TRUE}) and the Psi features.
\item \code{dummy_fit}: The fitted dummy coding model, which dummy-codes categorical
features.
For \code{apply_augmentation()}, a matrix containing the augmented data,
which includes both the raw features (if \code{include_raw = TRUE}) and the Psi
features. If \code{include_raw = FALSE}, only the Psi features are returned.
}
}
\description{
These helper functions augment the Psi matrix in NeRF+ with raw
features. Moreover, the raw features are dummy-coded if they are
categorical.
}
\keyword{internal}
