% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-embeddings.R
\name{oos_network_embedding}
\alias{oos_network_embedding}
\title{Out-of-sample network embedding method}
\usage{
oos_network_embedding(embedding_fit, A, A_full)
}
\arguments{
\item{embedding_fit}{Previously trained network embedding; should be
output of \code{fit_network_embedding()}.}

\item{A}{Adjacency matrix corresponding to the training data points.}

\item{A_full}{Adjacency matrix representing the full set of nodes (training +
testing nodes in that order).}
}
\value{
A list of two:
\itemize{
\item \code{X}: A list of matrices, each corresponding to a method in
\code{embedding_fit$method}, where each matrix has \code{ndim} columns representing
the embedding for the out-of-sample points.
\item \code{method}: The method(s) used for embedding.
}
}
\description{
This function takes in out-of-sample points (which were not available at
training time) and embeds them in a previously trained network embedding
space.
}
\examples{
data(example_data)
embedding_fit <- fit_network_embedding(
  example_data$A, ndim = 2, method = "laplacian"
)
oos_embedding <- oos_network_embedding(
  embedding_fit, A = example_data$A, A_full = example_data$A_full
)

}
