% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-app-modules.R
\name{plotUI}
\alias{plotUI}
\alias{plotServer}
\title{Basic plot module - UI}
\usage{
plotUI(id, border = FALSE, spinner = FALSE)

plotServer(
  id,
  plot_fun,
  plot_options = TRUE,
  modes = c("ggplot", "plotly"),
  border = FALSE,
  spinner = TRUE,
  error_msg = ""
)
}
\description{
Basic plot module - UI

Basic plot module - Server
}
\keyword{internal}
