% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnc-predict.R
\name{predict.rnc}
\alias{predict.rnc}
\title{Predict method for rnc (regression with network cohesion) objects}
\usage{
\method{predict}{rnc}(object, x, A_full, nodeids = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{rnc} containing the fitted model.}

\item{x}{A numeric matrix or data frame of predictors (features) to make
predictions for; size n x p.}

\item{A_full}{An adjacency matrix representing the network structure for
the full set of nodes (training + testing nodes in that order)}

\item{nodeids}{(Optional) vector of node IDs of length n.
If provided, node IDs indicate the rows of A_full, corresponding to each
sample. If not provided, the rows of A_full are assumed to be in the order
of (x_train, x).}

\item{...}{Not used.}
}
\value{
A list of two:
\itemize{
\item \code{y}: A numeric vector of predictions for the response variable.
\item \code{alpha}: A numeric vector of node-specific effects (intercepts).
}
}
\description{
Predict method for rnc (regression with network cohesion) objects
}
\examples{
data(example_data)

# demonstrate prediction for RNC fit
rnc_fit <- rnc(
  x = example_data$x, y = example_data$y, A = example_data$A,
  lambda_netcoh = 1, lambda_x = 0.5
)
predicted_out <- predict(
  rnc_fit, x = example_data$xtest, A_full = example_data$A_full
)

# demonstrate prediction for RNC CV fit
cv_fit <- rnc_cv(
  x = example_data$x, y = example_data$y, A = example_data$A,
  lambdas_netcoh = c(0, 0.1, 1),
  lambdas_x = c(0.1, 0.5)
)
predicted_out <- predict(
  cv_fit, x = example_data$xtest, A_full = example_data$A_full
)

}
