% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnc-predict.R
\name{predict_rnc_linear}
\alias{predict_rnc_linear}
\title{Predict method for rnc linear regression objects}
\usage{
predict_rnc_linear(object, x, A_full, nodeids = NULL)
}
\arguments{
\item{object}{An object of class \code{rnc} containing the fitted model.}

\item{x}{A numeric matrix or data frame of predictors (features) to make
predictions for; size n x p.}

\item{A_full}{An adjacency matrix representing the network structure for
the full set of nodes (training + testing nodes in that order)}

\item{nodeids}{(Optional) vector of node IDs of length n.
If provided, node IDs indicate the rows of A_full, corresponding to each
sample. If not provided, the rows of A_full are assumed to be in the order
of (x_train, x).}
}
\description{
Predict method for rnc linear regression objects
}
\keyword{internal}
