% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnc.R
\name{rnc_linear}
\alias{rnc_linear}
\title{Fit linear regression with network cohesion}
\usage{
rnc_linear(
  x,
  y,
  A,
  nodedegrees = NULL,
  nodeids = NULL,
  lambda_netcoh,
  lambda_x = 0,
  lambda_l = 0.05,
  low_dim = NULL
)
}
\arguments{
\item{x}{A numeric matrix or data frame of predictors (features); size n x p.
Should be centered so that each column has mean 0.}

\item{y}{A numeric vector of responses of length n. Should be centered so
that the mean is 0.}

\item{A}{An adjacency matrix representing the network structure.}

\item{nodedegrees}{(Optional) A vector of node degrees. If provided, it is
used to compute the graph Laplacian. If not provided, it is computed from
the adjacency matrix \code{A}.}

\item{nodeids}{(Optional) vector of node IDs of length n.
If provided, node IDs indicate the rows of A, corresponding to each
sample. If not provided, the rows of A are assumed to be in the same order
as the rows of x and y.}

\item{lambda_netcoh}{Regularization parameter for the network cohesion term.}

\item{lambda_x}{(Optional) Regularization parameter for the (linear)
covariates. Default is 0 (i.e., no regularization).}

\item{lambda_l}{(Optional) Regularization parameter for the graph Laplacian.
Default is 0.05.}

\item{low_dim}{(Optional) If \code{TRUE}, the algorithm will use a naive solver
for low-dimensional problems. Default is \code{NULL}, which will use the navie
low-dimensional solver if the number of covariates is <= 1/5 * the number
of samples. Only used if \code{family = "linear"}.}
}
\description{
Fit linear regression with network cohesion
}
\keyword{internal}
