% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnc-cv.R
\name{rnc_linear_path}
\alias{rnc_linear_path}
\title{Fit regularization path for linear regression with network cohesion}
\usage{
rnc_linear_path(
  x,
  y,
  L,
  x_test = NULL,
  y_test = NULL,
  L22 = NULL,
  L21 = NULL,
  lambda_grid,
  nodeids = NULL,
  nodeids_test = NULL,
  ...
)
}
\description{
Fit regularization path for linear regression with network cohesion
}
\keyword{internal}
