% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnc.R
\name{rnc_logistic}
\alias{rnc_logistic}
\title{Fit logistic regression with network cohesion}
\usage{
rnc_logistic(
  x,
  y,
  A,
  nodedegrees = NULL,
  nodeids = NULL,
  lambda_netcoh,
  lambda_x = 0,
  lambda_l = 0.05,
  init = NULL,
  newton_maxit = 50,
  newton_tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numeric matrix or data frame of predictors (features); size n x p.
Should be centered so that each column has mean 0.}

\item{y}{A numeric vector of responses of length n. Should be centered so
that the mean is 0.}

\item{A}{An adjacency matrix representing the network structure.}

\item{nodedegrees}{(Optional) A vector of node degrees. If provided, it is
used to compute the graph Laplacian. If not provided, it is computed from
the adjacency matrix \code{A}.}

\item{nodeids}{(Optional) vector of node IDs of length n.
If provided, node IDs indicate the rows of A, corresponding to each
sample. If not provided, the rows of A are assumed to be in the same order
as the rows of x and y.}

\item{lambda_netcoh}{Regularization parameter for the network cohesion term.}

\item{lambda_x}{(Optional) Regularization parameter for the (linear)
covariates. Default is 0 (i.e., no regularization).}

\item{lambda_l}{(Optional) Regularization parameter for the graph Laplacian.
Default is 0.05.}

\item{init}{(Optional) initial values for the optimization algorithm to fit
logistic regression. Ignored for linear regression.}

\item{newton_maxit}{Maximum number of Newton iterations when fitting
logistic regression. Default is 50. Ignored for linear regression.}

\item{newton_tol}{Tolerance for convergence of Newton iterations when fitting
logistic regression. Default is 1e-4. Ignored for linear regression.}

\item{verbose}{Logical indicating whether to print progress messages.}
}
\description{
Fit logistic regression with network cohesion
}
\keyword{internal}
