% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnc.R
\name{rnc_logistic_scaled_laplacian}
\alias{rnc_logistic_scaled_laplacian}
\title{Fit scaled logistic regression with a precomputed network Laplacian}
\usage{
rnc_logistic_scaled_laplacian(
  x,
  y,
  L,
  lambda_netcoh,
  lambda_x = 0,
  init = NULL,
  newton_maxit = 50,
  newton_tol = 1e-04,
  verbose = FALSE
)
}
\description{
Fit scaled logistic regression with a precomputed network Laplacian
}
\keyword{internal}
