% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{rsq_narm_vec}
\alias{rsq_narm_vec}
\title{Evaluate r-squared with NA handling}
\usage{
rsq_narm_vec(truth, estimate, ...)
}
\arguments{
\item{truth}{A numeric vector of true values.}

\item{estimate}{A numeric vector of estimated values.}

\item{...}{Additional arguments passed to \code{\link[yardstick:rsq_vec]{yardstick::rsq_vec()}}}
}
\value{
A numeric value representing the R-squared value, or 0 if the
estimate is a constant vector.
}
\description{
This function computes the R-squared value and returns 0
(instead of NA) if the estimate is a constant vector.
}
\keyword{internal}
