% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny NeRF+ Interpret application}
\usage{
run_app(
  data_list = NULL,
  object = NULL,
  fi_results = NULL,
  loo_results = NULL,
  max_request_size = 100 * 1024^2,
  ...
)
}
\arguments{
\item{data_list}{A list containing the data used for interpretation. Should
have the same format as the \code{data_list} output from \code{\link[=interpret_nerfplus]{interpret_nerfplus()}}.}

\item{object}{The fitted NeRF+ model. Should have the same format as the
\code{object} output from \code{\link[=interpret_nerfplus]{interpret_nerfplus()}}.}

\item{fi_results}{A list containing the feature importance results. Should
have the same format as the \code{fi_results} output from \code{\link[=interpret_nerfplus]{interpret_nerfplus()}}.}

\item{loo_results}{A list containing the sample influence results. Should
have the same format as the \code{loo_results} output from \code{\link[=interpret_nerfplus]{interpret_nerfplus()}}.}

\item{max_request_size}{Maximum request size for file uploads. Defaults to
100 MB.}

\item{...}{Additional arguments to pass to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
A shiny application
}
\description{
Run Shiny NeRF+ Interpret application
}
\examples{
\donttest{
if (interactive()) {
  # launch app
  run_app()
}

# or run with arguments
data(example_data)
nerfplus_out <- nerfplus(
  x = example_data$x, y = example_data$y, A = example_data$A,
  lambda_netcoh = 1,
  lambda_embed = 0.1,
  lambda_raw = 2,
  lambda_stump = 3,
  family = "linear", embedding = "laplacian", sample_split = "none"
)
out <- interpret_nerfplus(
  nerfplus_out, x = example_data$x, y = example_data$y, A = example_data$A,
  xtest = example_data$xtest, ytest = example_data$ytest,
  A_full = example_data$A_full
)

if (interactive()) {
  run_app(
    data_list = out$data_list,
    object = out$object,
    fi_results = out$fi_results,
    loo_results = out$loo_results
  )
}
}

}
