% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-app-widgets.R
\name{run_button}
\alias{run_button}
\title{Run button}
\usage{
run_button(inputId = "run_methods", label = "Run methods", ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{...}{Additional arguments to pass to \code{shiny::actionButton()}.}
}
\description{
Customized action button using \code{actionButton()}.
}
\keyword{internal}
