% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{shared_args_util}
\alias{shared_args_util}
\title{Shared arguments for NeRF+ utility functions}
\arguments{
\item{tree_info}{Output of \code{\link[ranger:treeInfo]{ranger::treeInfo()}} for a single tree.}

\item{tree_infos}{List of size ntrees with each entry being the output of
\code{\link[ranger:treeInfo]{ranger::treeInfo()}}.}

\item{tree_paths}{List of size ntrees with paths for each tree; typically the
output of \code{\link[=get_forest_paths]{get_forest_paths()}}.}

\item{node_preds}{Matrix of terminal node predictions for each tree in the
random forest; typically, output from
\code{predict(rf_fit, x, type = "terminalNodes")$predictions}, where
\code{rf_fit} is a fitted \code{ranger} object and \code{x} is the data frame or matrix.}

\item{unordered_factors}{Vector of column names corresponding to unordered
factor variables in the data.}

\item{inbag_counts}{List of size ntrees with inbag counts for each tree;
typically the output from \code{rf_fit$inbag.counts}, where \code{rf_fit} is a
fitted \code{ranger} object. Ignored if \code{normalize = FALSE}.}

\item{include_raw}{Logical indicating whether to include the raw covariates
in the NeRF+ model. Default is \code{TRUE}.}
}
\description{
Shared arguments for NeRF+ utility functions
}
\keyword{internal}
