% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-app-modules.R
\name{tableServer}
\alias{tableServer}
\title{Basic table module - Server}
\usage{
tableServer(
  id,
  table_fun,
  table_options = TRUE,
  mode = c("DT", "kable"),
  caption = NULL,
  border = FALSE,
  spinner = TRUE,
  error_msg = "",
  ...
)
}
\description{
Basic table module - Server
}
\keyword{internal}
