% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{tree_local_fi}
\alias{tree_local_fi}
\title{Compute local feature importance for a tree in NeRF+}
\usage{
tree_local_fi(
  tree_object,
  x,
  A_full = NULL,
  nodeids = NULL,
  x_means = NULL,
  grouped_features = NULL,
  L_full = NULL,
  solve_cache = NULL
)
}
\arguments{
\item{tree_object}{A fitted tree object from NeRF+.}

\item{x}{A data frame or matrix used to compute the feature importances.}

\item{A_full}{An adjacency matrix representing the network structure for
the full set of nodes (training + testing nodes in that order)}

\item{nodeids}{(Optional) vector of node IDs of length n.
If provided, node IDs indicate the rows of A, corresponding to each
sample. If not provided, the rows of A are assumed to be in the same order
as the rows of x and y.}

\item{x_means}{A named vector of means for each feature in \code{x}. If \code{NULL},
the means are computed from \code{x}.}

\item{grouped_features}{A list of features to group together for permutation
importance. Each element of the list should be a character vector of
feature names. If \code{NULL}, each feature is treated as a separate group.
Typically, this is the output of \code{get_grouped_tree_features()}.}
}
\value{
A tibble with n rows and p columns, where n is the number of samples
and p is the number of features. Each column corresponds to a feature,
and each row corresponds to a sample. The values represent the local
feature importance scores for each feature and sample.
}
\description{
Compute local feature importance for a tree in NeRF+
}
\keyword{internal}
