% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{tree_permute_fi}
\alias{tree_permute_fi}
\title{Compute permutation importance for a tree in NeRF+}
\usage{
tree_permute_fi(
  tree_object,
  x,
  y,
  A_full = NULL,
  nodeids = NULL,
  metric,
  grouped_features = NULL,
  B = 10,
  return_preds = FALSE,
  L_full = NULL,
  solve_cache = NULL
)
}
\arguments{
\item{tree_object}{A fitted tree object from NeRF+.}

\item{x}{A data frame or matrix used to compute the feature importances.}

\item{y}{A vector of responses.}

\item{A_full}{An adjacency matrix representing the network structure for
the full set of nodes (training + testing nodes in that order)}

\item{nodeids}{(Optional) vector of node IDs of length n.
If provided, node IDs indicate the rows of A, corresponding to each
sample. If not provided, the rows of A are assumed to be in the same order
as the rows of x and y.}

\item{metric}{A function to compute the metric used for global feature
importances. Defaults to R-squared for regression and AUROC for
classification. Ignored if \code{method} is \code{"local"}.}

\item{grouped_features}{A list of features to group together for permutation
importance. Each element of the list should be a character vector of
feature names. If \code{NULL}, each feature is treated as a separate group.
Typically, this is the output of \code{get_grouped_tree_features()}.}

\item{B}{Number of bootstrap samples for permutation-based feature
importance. Alternatively, a list of permutation indices can be
passed to \code{B}, where each element is a vector of indices for permuting
features.}

\item{return_preds}{Logical indicating whether to return the predictions
for each permutation. If \code{TRUE}, the function returns a list of predictions
for each permutation. If \code{FALSE}, it returns the computed feature
importance scores.}
}
\value{
If \code{return_preds} is \code{FALSE}, a tibble with two columns:
\itemize{
\item \code{var}: The name of the feature.
\item \code{importance}: The computed feature importance score.
If \code{return_preds} is \code{TRUE}, a list where each element corresponds to a
feature group and contains a list of predictions for each permutation.
}
}
\description{
Compute permutation importance for a tree in NeRF+
}
\keyword{internal}
