% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nerfplus.R
\name{nerfplus_psi}
\alias{nerfplus_psi}
\alias{fit_psi}
\alias{apply_psi}
\title{Get the Psi matrix in NeRF+ containing decision stump features}
\usage{
fit_psi(
  x,
  tree_info,
  tree_paths,
  node_preds,
  unordered_factors = NULL,
  normalize = FALSE,
  inbag_counts = NULL
)

apply_psi(
  x,
  tree_info,
  tree_paths,
  node_preds,
  unordered_factors = NULL,
  psi_unique_values = NULL
)
}
\arguments{
\item{x}{A data frame or matrix containing the data to feed through decision
tree to obtain the Psi (i.e., decision stump features) matrix.}

\item{tree_info}{Output of \code{\link[ranger:treeInfo]{ranger::treeInfo()}} for a single tree.}

\item{tree_paths}{List of size ntrees with paths for each tree; typically the
output of \code{\link[=get_forest_paths]{get_forest_paths()}}.}

\item{node_preds}{Matrix of terminal node predictions for each tree in the
random forest; typically, output from
\code{predict(rf_fit, x, type = "terminalNodes")$predictions}, where
\code{rf_fit} is a fitted \code{ranger} object and \code{x} is the data frame or matrix.}

\item{unordered_factors}{Vector of column names corresponding to unordered
factor variables in the data.}

\item{normalize}{Logical indicating whether to normalize the Psi matrix by
number of training samples in each child node. Defaults to \code{FALSE}.}

\item{inbag_counts}{List of size ntrees with inbag counts for each tree;
typically the output from \code{rf_fit$inbag.counts}, where \code{rf_fit} is a
fitted \code{ranger} object. Ignored if \code{normalize = FALSE}.}

\item{psi_unique_values}{A named list of unique values for each Psi feature,
typically obtained from the output of \code{fit_psi()}.}
}
\value{
For \code{fit_psi()}, a list of two:
\itemize{
\item \code{psi}: A data frame containing the Psi matrix, where each column
corresponds to a decision stump feature
\item \code{psi_unique_values}: A named list of unique values for each Psi feature
(only returned if \code{normalize = TRUE}).
For \code{apply_psi()}, a data frame containing the Psi matrix, where each
column corresponds to a decision stump feature.
}
}
\description{
These helper functions fit or apply a previously fitted
Psi mapping (i.e., decision stump feature mapping) to data.
}
\keyword{internal}
