% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnc-cv.R
\name{rnc_logistic_path}
\alias{rnc_logistic_path}
\title{Fit regularization path for logistic regression with network cohesion}
\usage{
rnc_logistic_path(
  x,
  y,
  A,
  x_test = NULL,
  y_test = NULL,
  A_full = NULL,
  lambda_grid,
  nodeids = NULL,
  nodeids_test = NULL,
  newton_maxit = 50,
  newton_tol = 1e-04,
  ...
)
}
\description{
Fit regularization path for logistic regression with network cohesion
}
\keyword{internal}
