% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{tree_mdiplus_fi}
\alias{tree_mdiplus_fi}
\title{Compute MDI+ feature importance for a tree in NeRF+}
\usage{
tree_mdiplus_fi(
  tree_object,
  x,
  y,
  A_full = NULL,
  nodeids = NULL,
  metric,
  x_means = NULL,
  grouped_features = NULL
)
}
\arguments{
\item{tree_object}{A fitted tree object from NeRF+.}

\item{x}{A data frame or matrix used to compute the feature importances.}

\item{y}{A vector of responses.}

\item{A_full}{An adjacency matrix representing the network structure for
the full set of nodes (training + testing nodes in that order)}

\item{nodeids}{(Optional) vector of node IDs of length n.
If provided, node IDs indicate the rows of A, corresponding to each
sample. If not provided, the rows of A are assumed to be in the same order
as the rows of x and y.}

\item{metric}{A function to compute the metric used for global feature
importances. Defaults to R-squared for regression and AUROC for
classification. Ignored if \code{method} is \code{"local"}.}

\item{x_means}{A named vector of means for each feature in \code{x}. If \code{NULL},
the means are computed from \code{x}.}

\item{grouped_features}{A list of features to group together for permutation
importance. Each element of the list should be a character vector of
feature names. If \code{NULL}, each feature is treated as a separate group.
Typically, this is the output of \code{get_grouped_tree_features()}.}
}
\value{
A tibble with two columns:
\itemize{
\item \code{var}: The name of the feature.
\item \code{importance}: The computed feature importance score.
}
}
\description{
Compute MDI+ feature importance for a tree in NeRF+
}
\keyword{internal}
